Added download link to Readme
I've seen people asking for a download link to the precompiled binaries. There's no obvious way to get to the releases page. Therefore, I download section in the main readme file so that users, especially those who are not developers and are not experienced with using GitHub can easily find the link to download the latest version.