App Version Updated to Version 1804
The application version has been updated from version 10.0.0.0 to version 10.0.1804.0. I've not updated the major or minor version; so it's still version 10.0 as of now. The 1804 signifies that it is the April 2018 code.

Since there is no automated versioning system implimented yet, this allows us to at least easily update the version easily without worrying about whether to increment the major or minor version.

Note: I originally wanted to go with something like 180418, (YYMMDD) to include the day. This way, whoever commits a code, they would just update the version number to reflect the date. That way, it'd be super easy to tell which "build" one is runnign for troubleshooting reasons. But I went with just YYMM (1804) because we'd only need to increment the version at least once a month.

But I'm interested to hear what you think of how versioning should be done.