Resolves User Interface Issue Identified in #83
This commit resolves an issue that was identified in #83 by @Anixx.

The window size for the "Confirmation" dialog box's height value is too small, causing an overlapping issue with the "group" control. This commit increases the window height by one pixel that fixes this visual issue. This fix was suggested by @peterooch and verified working. So credit goes to them.