Increased quality of icons used in the app
Icons outlines have been made darker, returning it back to mostly the way they used to look in Original_plus version and earlier.

Note, however, that the icons are not a complete revert to the old look. It's an evolution of the latest bitmaps used in the master branch to increase sharpness and clarity of the icons. Additionally, the directory tree line color has not been reverted to the old look intentionally.