Complete Port of Original Help Docs to Fully Functional, interactive, HTML
The original documentations for winfile, as extracted from the Windows 98 operating system, has now been ported to an open, and fully functional web based help system, complete with hyper-link support and table of contents.

This represents the initial milestone "freeing" the documentation, making it easy to read on any device. And now that this is done, contributors can now easily add information to reflect the current state of the applicaion.

Furthermore, I'm hoping to have the webpage that is stored on the reposity serve as a live, up-to-date, documentation for the applicaion, that can be launched right from the help menu. And, and the same time, with subsequent releases of Winfile, a forzen version of these documents can be distributed with the binaries, allowing that particular release enjoy a fully offline copy of the help files, as they were at the time of the release. This can be especially useful in older computers without internet access.

I have not done extensive testing in all the differnt browsers, and, but it appears all desktop browsers are able to view the files perfectly fine. I have yet to test mobile browser compatibilities.

But that's where the community comes in. By having the help file out there and easily accessible, it can be improved quickly whereas, before, the documentation was trapped in propriotary file fomates that were difficult to read.

Best