Updated Readme file to include link to #88 (#117)
The issue thread #88 offers some great points on what makes a good pull request. However, it is buried under a lot of different issues and can be difficult to reach, especially for new contributors who might not know this discussion exists. As such, I felt it necessary to add a direct link to the thread under the "Contributing" section of the readme file.