Assume you have the following files in a directory
schin.jpg
sec1.bla
sec2.jpg
sh.jpg
sim.txt
ssh.ext

If you start typing in the right dir pane with 's' 's' 'h' the selection lands on sh.jpg instead of ssh.ext
This means if the first two characters are the same, the second is swallowed, and it matches sh.jpg.
This is a bit od, because if you rely on typing you end up on the wrong filename.