Took over this one too. Took me the longest to check. bDoMoveRename is really used that late, and since we have the check for FUNC_COPY this is more logical. // if (bSameFile && pCopyInfo->dwFunc == FUNC_COPY Checking the path is also good for all pathes // IsInvalidPath ...
I was a bit struggling with the removal of the switch, because as mentiond earlier there might be FUNC_HARD, FUNC_LINK ... and so on
Anyhow the 'switch' will then return, also the 'if (bSameFile ....' will become a little more clumsy, because all affected FUNC_ are mentioned

I guess this was the last commit. Thx for digging into and providing leaner sections!